chore: reorganize queries
